chore: make workspace clippy-clean
Ensure `cargo clippy --workspace --all-targets --all-features -- -D warnings` passes by addressing clippy lints across core crates and test targets.

Updates include replacing manual arithmetic helpers (is_multiple_of/div_ceil/clamp), removing identity ops/unnecessary casts, and cleaning up test/bench helpers and config initialization patterns.

+++ crates/aero-cpu-core/src/interp/tier0/ops_atomic.rs
@@ -112,10 +112,8 @@ fn exec_cmpxchg8b<B: CpuBus>(
     }
 
     let addr = calc_ea(state, instr, next_ip, true)?;
-    let expected =
-        (state.read_reg(Register::EDX) << 32) | (state.read_reg(Register::EAX) as u32 as u64);
-    let replacement =
-        (state.read_reg(Register::ECX) << 32) | (state.read_reg(Register::EBX) as u32 as u64);
+    let expected = (state.read_reg(Register::EDX) << 32) | state.read_reg(Register::EAX);
+    let replacement = (state.read_reg(Register::ECX) << 32) | state.read_reg(Register::EBX);
     let lock = instr.has_lock_prefix();
 
     let (old, swapped) = if lock {
